llama2/requirements.txt
fastapi>=0.99.1
uvicorn>=0.22.0
pydantic>=1.10.10
celery>=5.3.1
redis>=4.6.0
python-dotenv>=1.0.0
transformers>=4.32.0
torch>=2.0.1
accelerate>=0.21.0
bitsandbytes>=0.41.0
scipy>=1.11.1
optimum>=1.12.0
auto-gptq>=0.7.1

#Some of the dependencies are changed for incompatible versions